5.31 yj提交3

+++ pharmacy_common/src/main/java/com/kgc/entity/CgrkOrder.java
@@ -47,7 +47,7 @@ public class CgrkOrder implements Serializable {
      */
     @TableField("medicinePrice")
     @ExcelFiled("库存总价值")
-    private Double medicineprice;
+    private BigDecimal medicineprice;
 
     /**
      * 采购单Id
@@ -141,7 +141,6 @@ public class CgrkOrder implements Serializable {
     @TableField("updateTime")
     @ExcelFiled("修改时间")
     @JsonFormat(locale = "zh", timezone = "GMT+8", pattern = "yyyy-MM-dd mm:hh:ss")
-
     private Date updateTime;
 
     /**
